ai-moderation/requirements.txt
# AI/ML Libraries
torch==2.0.1
torchvision==0.15.2
transformers==4.32.0
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76

# Content Moderation
nudenet==2.0.9

# Pose Detection (MediaPipe alternative that's lighter)
mediapipe==0.10.3

# Image Understanding
sentence-transformers==2.2.2

# Web Framework
flask==2.3.2
flask-cors==4.0.0

# AWS/Queue Integration
boto3==1.28.25
requests==2.31.0

# Database
mysql-connector-python==8.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.1.1